Tests AdductProcessing
